backend/requirements.txt
asgiref==3.11.1
cryptography==46.0.1
Django==6.0.3
djangorestframework==3.17.0
psycopg[binary]==3.2.10
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
